requirements.txt
Flask==1.1.2
Flask-RESTful==0.3.8
numpy==1.18.1
matplotlib==3.2.0
pandas==1.0.3
pyyaml==5.3.1
scikit-learn==0.22.2.post1
